requirements.txt
discord.py>=2.0.0
aiosqlite
python-dotenv
requests
scrython